Automated deploys to Elastic Beanstalk
This reworks the project and adds supporting files for automated deploys to Elastic Beanstalk.

The underlying principle here is that I can run `eb deploy <env>` on a fresh environment and have a new copy of the application up and running. I can also deploy to existing environments without extra configuration.

Flask now just serves API endpoints, while the output of `yarn build` in our create-react-app folder, frontend, is copied into the Apache DocumentRoot.  Flask is mounted in Apache at the `/api` path, which required some changes to the routing.

This has to fight Elastic Beanstalk's default python configuration a little bit. For example, in addition to hosting the static part of the website in the DocumentRoot, we also have to override the default Alias EB provides for static files.

The three complex part of this build are
* numpy and scipy have to be built from source, _before_ pandas and sklearn. This requires some gnarly and poorly documented packages to be installed.
* `yarn build` happens on the server, so this installs a fulle node.js environment.
* make Flask and our staticfiles co-exist and avoid serving 'index.html' and other root level resources from Flask

This also adds a setup.py which lets create more flexible reproducible builds.